git-hook: fix enabling/disabling
This script originally used just chmod but that broke at some point with
git now complaining about hooks that aren't executable. The mv .disabled
bit was added but not propertly tested before committing, so enabling a
hook didn't work because we only checked for enbled hook file names.

Fix this and drop the now pointless chmod.